# shared/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Team4s v3 API
  version: 0.1.0
  description: |
    Baseline OpenAPI contract for current endpoints, including admin content-management routes.
  license:
    name: Proprietary
    url: https://team4s.dev/license
servers:
  - url: /
    description: Relative API base
...
security: []
paths:
...
  /api/v1/fansubs:
    get:
      tags: [Fansubs]
      summary: List fansub groups
      operationId: listFansubGroups
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/FansubStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 1
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 24
      responses:
        "200":
          description: Fansub group list payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FansubGroupListResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Fansubs]
      summary: Create fansub group
      operationId: createFansubGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FansubGroupCreateRequest"
      responses:
        "201":
          description: Created fansub group payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FansubGroupResponse"
        "400":
          description: Invalid request body or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate slug or name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/fansubs/{id}:
    get:
      tags: [Fansubs]
      summary: Get fansub group by id
      operationId: getFansubGroupById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Fansub group payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FansubGroupResponse"
        "400":
          description: Invalid fansub id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Fansub group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Fansubs]
      summary: Update fansub group
      operationId: updateFansubGroup
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FansubGroupPatchRequest"
      responses:
        "200":
          description: Updated fansub group payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FansubGroupResponse"
        "400":
          description: Invalid request body or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Fansub group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate slug or name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Fansubs]
      summary: Delete fansub group
      operationId: deleteFansubGroup
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Fansub group deleted
        "400":
          description: Invalid fansub id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Fansub group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/anime/{animeId}/fansubs:
    get:
      tags: [Fansubs]
      summary: List fansub relations for an anime
      operationId: listAnimeFansubs
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Anime fansub relation list payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeFansubListResponse"
        "400":
          description: Invalid anime id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Anime not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/anime/{animeId}/fansubs/{fansubId}:
    post:
      tags: [Fansubs]
      summary: Attach fansub group to anime
      operationId: attachAnimeFansub
      security:
        - bearerAuth: []
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: fansubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimeFansubAttachRequest"
      responses:
        "201":
          description: Anime fansub relation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeFansubResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Anime or fansub group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Relation already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Fansubs]
      summary: Detach fansub group from anime
      operationId: detachAnimeFansub
      security:
        - bearerAuth: []
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: fansubId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Relation deleted
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin permission required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Relation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SignedToken
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
...
    PaginationMeta:
      type: object
      required: [total, page, per_page, total_pages]
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        per_page:
          type: integer
          format: int32
        total_pages:
          type: integer
          format: int32
...
    FansubStatus:
      type: string
      enum: [active, inactive, dissolved]
    FansubGroupSummary:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        name:
          type: string
        logo_url:
          type: string
          nullable: true
    FansubGroup:
      type: object
      required:
        - id
        - slug
        - name
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
          maxLength: 120
        name:
          type: string
          maxLength: 120
        description:
          type: string
          nullable: true
        history:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        banner_url:
          type: string
          nullable: true
        founded_year:
          type: integer
          format: int32
          nullable: true
        dissolved_year:
          type: integer
          format: int32
          nullable: true
        status:
          $ref: "#/components/schemas/FansubStatus"
        website_url:
          type: string
          nullable: true
        discord_url:
          type: string
          nullable: true
        irc_url:
          type: string
          nullable: true
        country:
          type: string
          maxLength: 80
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AnimeFansubRelation:
      type: object
      required:
        - anime_id
        - fansub_group_id
        - is_primary
        - created_at
      properties:
        anime_id:
          type: integer
          format: int64
        fansub_group_id:
          type: integer
          format: int64
        is_primary:
          type: boolean
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        fansub_group:
          $ref: "#/components/schemas/FansubGroupSummary"
    FansubGroupCreateRequest:
      type: object
      required: [slug, name, status]
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 120
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
        history:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        banner_url:
          type: string
          nullable: true
        founded_year:
          type: integer
          format: int32
          nullable: true
        dissolved_year:
          type: integer
          format: int32
          nullable: true
        status:
          $ref: "#/components/schemas/FansubStatus"
        website_url:
          type: string
          nullable: true
        discord_url:
          type: string
          nullable: true
        irc_url:
          type: string
          nullable: true
        country:
          type: string
          maxLength: 80
          nullable: true
    FansubGroupPatchRequest:
      type: object
      minProperties: 1
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 120
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
        history:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        banner_url:
          type: string
          nullable: true
        founded_year:
          type: integer
          format: int32
          nullable: true
        dissolved_year:
          type: integer
          format: int32
          nullable: true
        status:
          $ref: "#/components/schemas/FansubStatus"
        website_url:
          type: string
          nullable: true
        discord_url:
          type: string
          nullable: true
        irc_url:
          type: string
          nullable: true
        country:
          type: string
          maxLength: 80
          nullable: true
...
    AnimeFansubAttachRequest:
      type: object
      properties:
        is_primary:
          type: boolean
          default: false
        notes:
          type: string
          nullable: true
    FansubGroupListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FansubGroup"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    FansubGroupResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/FansubGroup"
...
    AnimeFansubListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AnimeFansubRelation"
    AnimeFansubResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AnimeFansubRelation"
...